pyproject.toml
[tool.poetry]
name = "orator"
version = "0.9.9"
description = "The Orator ORM provides a simple yet beautiful ActiveRecord implementation."

license = "MIT"

authors = [
    "Sébastien Eustace <sebastien@eustace.io>"
]

readme = 'README.rst'

repository = "https://github.com/sdispater/orator"
homepage = "https://orator-orm.com/"

keywords = ['database', 'orm']


[tool.poetry.dependencies]
python = "^3.6"
backpack = "^0.1"
blinker = "^1.4"
cleo = "^0.8.1"
inflection = "^0.3"
Faker = "^0.8"
lazy-object-proxy = "^1.2"
pendulum = "^2.0.5"
pyaml = "^16.12"
pyyaml = "^5.1"
Pygments = "^2.2"
simplejson = "^3.10"
six = "^1.10"
wrapt = "^1.10"

# Extras
psycopg2 = { version = "^2.7", optional = true }
PyMySQL = { version = "^0.7", optional = true }
mysqlclient = { version = "^1.3", optional = true }


[tool.poetry.dev-dependencies]
flexmock = "0.9.7"
pytest = "^3.5"
pytest-mock = "^1.6"
tox = "^3.5"
pre-commit = "^1.11"
black = "^20.8b1"
isort = "^5.6.4"


[tool.poetry.extras]
mysql = ["mysqlclient"]
mysql-python = ["PyMySQL"]
pgsql = ["psycopg2"]


[tool.poetry.scripts]
orator = 'orator.commands.application:application.run'


[build-system]
requires = ["poetry>=0.12a3"]
build-backend = "poetry.masonry.api"
